functions/imageModeration/requirements.txt
Pillow==9.4.0
blurhash-python==1.2.0